+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Add mapping profiles deleting. UIDEXP-133.
 * Implement record type filter behavior in add mapping profile transformations modal. UIDEXP-69.
 * Accommodate `FullScreenView` component usage according to [changes in stripes-data-transfer-components](https://github.com/folio-org/stripes-data-transfer-components/pull/88/files). UIDATIMP-535.
+* Implement search feature in mapping profile transformations. UIDEXP-130.
 
 ## [2.0.1](https://github.com/folio-org/ui-data-export/tree/v2.0.1) (2020-07-09)
 [Full Changelog](https://github.com/folio-org/ui-data-export/tree/v2.0.0...v2.0.1)